+++ packages/EasyErrorHandler/config/ErrorHandlerCompilerPass.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace EonX\EasyErrorHandler\DependencyInjection\Compiler;
+namespace EonX\EasyErrorHandler\Config;
 
 use EonX\EasyErrorHandler\Bridge\Symfony\Interfaces\TraceableErrorHandlerInterface;
 use EonX\EasyErrorHandler\Bridge\Symfony\TraceableErrorHandler;
