+++ packages/EasyErrorHandler/docs/install.md
@@ -26,7 +26,7 @@ yourself:
 return [
     // Other bundles ...
 
-    EonX\EasyErrorHandler\EasyErrorHandlerSymfonyBundle::class => ['all' => true],
+    EonX\EasyErrorHandler\Config\EasyErrorHandlerSymfonyBundle::class => ['all' => true],
 ];
 ```
