+++ packages/EasyErrorHandler/tests/Bridge/Symfony/Fixtures/app/config/bundles.php
@@ -3,7 +3,7 @@ declare(strict_types=1);
 
 use ApiPlatform\Symfony\Bundle\ApiPlatformBundle;
 use Doctrine\Bundle\DoctrineBundle\DoctrineBundle;
-use EonX\EasyErrorHandler\EasyErrorHandlerSymfonyBundle;
+use EonX\EasyErrorHandler\Config\EasyErrorHandlerSymfonyBundle;
 use EonX\EasyLogging\Bridge\Symfony\EasyLoggingSymfonyBundle;
 use Symfony\Bundle\FrameworkBundle\FrameworkBundle;
